grupo10/api/serializers.py
from rest_framework import serializers
from web.models import Product

class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = ['name', 'price', 'category', 'description']